range_diversity_analysis: diversity analysis

diversity_range_analysis biodiversity indices related to diversity-range plots

diversity_range_analysis(
  pam,
  xy_mat = NULL,
  lower_interval = 0.05,
  upper_interval = 0.95,
  raster_templete = NULL,
  niter = 100,
  return_null_dfield = FALSE,
  parallel = TRUE,
  n_cores = 2
)

Arguments

pam

A Presence-Absence-Matrix of matrix class or sparse matrix.

xy_mat

A two dimensional matrix with longitude and latitude data.

lower_interval

Lower interval.

upper_interval

Upper interval.

raster_templete

Araster templete.

niter

Number of iterations to obtain the distribution.

return_null_dfield

If TRUE the null distribution of dispersal field will be returned.

parallel

If TRUE the computations will be performed in parallel.

n_cores

Number of cores for the parallel computation.

Examples

if (FALSE) {
set.seed(111)
pam <- matrix(rbinom(10000,1,0.5),nrow = 100,ncol = 1000)
rdivan <- bamm::diversity_range_analysis(pam=pam,parallel = FALSE,
                                   return_null_dfield=TRUE)
bamm::plot(rdivan,plot_type="diversity_range")
# Lagomorphos
lagos_path <- system.file("extdata/conejos",
                          package = "bamm")
enm_path <- list.files(lagos_path,
                       pattern = ".tif",
                       full.names = TRUE)
en_models <- raster::stack(enm_path) >0.01
nonas <- which(!is.na(en_models[[1]][]))
xy_mat <- sp::coordinates(en_models[[1]])[ nonas,]
pam <- bamm::models2pam(en_models,sparse=FALSE)
rdivan <- bamm::diversity_range_analysis(pam=pam,parallel = FALSE,
                                   xy_mat=xy_mat,
                                   raster_templete = en_models[[1]],
                                   return_null_dfield=TRUE)
bamm::plot(rdivan,plot_type="diversity_range")
bamm::plot(rdivan,plot_type="diversity_range_map")
bamm::plot(rdivan,plot_type="diversity_range_interactive")

}